requirements.txt
boto3
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity